タスク11の提出をします。

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -11,8 +11,10 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentalsService;
 
 /**
  * 削除コントローラー
@@ -22,6 +24,8 @@ public class DeleteBookController {
 	final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
 	@Autowired
 	private BooksService booksService;
+	@Autowired
+	private RentalsService rentalsService;
 
 	/**
 	 * 対象書籍を削除する
@@ -33,12 +37,17 @@ public class DeleteBookController {
 	 */
 	@Transactional
 	@RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
-	public String deleteBook(Locale locale, @RequestParam("bookId") Integer bookId, Model model) {
+	public String deleteBook(Locale locale, @RequestParam("bookId") Integer bookId, Model model, RedirectAttributes redirectAttributes) {
 		logger.info("Welcome delete! The client locale is {}.", locale);
-		booksService.deleteBook(bookId);
-		model.addAttribute("bookList", booksService.getBookList());
-		return "home";
-
+		
+		boolean bookIdExists = rentalsService.selectBookInfo(bookId);
+		if (bookIdExists) {
+			redirectAttributes.addFlashAttribute("errorRentals", "現在、貸出し中のため削除できません。");
+			return "redirect:/details?bookId=" + bookId;
+		} else {
+			booksService.deleteBook(bookId);
+			return "redirect:/home";
+		}
 	}
 
 }
